+++ R/compile.R
@@ -1,18 +1,61 @@
 
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Retrieve the default CFLAGS R uses to compile shared libraries
+#' @return character string of default CFLAGS
+#' @examples
+#' cflags_default()
+#' @export
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+cflags_default <- function() {
+  system("R CMD CONFIG CFLAGS", intern = TRUE)
+}
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# Run code with the given R_MAKEVARS_USER setting
+# Similar idea to 'withr::with_makevars()'
+#
+# The reason this is done with "R_MAKEVARS_USER + CFLAGS" and not 
+# "Makevars + PKG_CFLAGS" is that if PKG_CFLAGS 
+# come **before** R's default flags i.e.
+#    clang  <PKG_CFLAGS>  <R default cflags>  file.c
+#
+# but 'CFLAGS' **replaces** R's default CFLAGS, i.e.
+#    clang <CFLAGS> file.c
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+with_makevars_file_env <- function(R_MAKEVARS_USER, code) {
+  default <- Sys.getenv('R_MAKEVARS_USER')
+  if (default == "") {
+    on.exit(Sys.unsetenv('R_MAKEVARS_USER'))
+  } else {
+    on.exit(Sys.setenv(R_MAKEVARS_USER = default))
+  }
+  Sys.setenv(R_MAKEVARS_USER = R_MAKEVARS_USER)
+  force(code)
+}
+
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #' Compile C code and load into R for use with \code{.Call()}
 #' 
 #' See also \code{?SHLIB}
 #' 
 #' @param code C code following the \code{.Call()} conventions.  Must
 #'        also include any \code{#include} statements.
+#' @param CFLAGS character string of flags for the C compiler. e.g. "-O3"
+#'        Default: NULL.  If specified this value will \emph{replace} the
+#'        default \code{CFLAGS} R would normally use.  To see these default
+#'        flags use \code{cflags_default()}.
 #' @param PKG_CPPFLAGS character string of flags for the C pre-processor.
+#'        Flags such as "-I", "-D" and "-U" go here.
 #'        Default: NULL
 #'        e.g. \code{PKG_CPPFLAGS = "-I/opt/homebrew/include"} to add the include path 
 #'        for homebrew to the compilation step. 
-#' @param PKG_LDFLAGS character string of flags when linking. Default: NULL.
-#'        e.g. \code{PKG_LDFLAGS = "-L/opt/homebrew/lib -lzstd"} to include the homebrew 
+#' @param PKG_LIBS character string of flags when linking. "-L" and "-l" flags
+#'        go here. Default: NULL.
+#'        e.g. \code{PKG_LIBS = "-L/opt/homebrew/lib -lzstd"} to include the homebrew 
 #'        libraries in the linker search path and to link to the \code{zstd}
 #'        library installed there. 
 #' @param env environment into which to assign the R wrapper functions.
@@ -72,8 +115,8 @@
 #' mul(99.5, 0.5)
 #' new_sqrt(c(1, 10, 100, 1000))
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-compile <- function(code, PKG_CPPFLAGS = NULL, PKG_LDFLAGS = NULL, env = parent.frame(), 
-                   overwrite = "callme", verbosity = 0) {
+compile <- function(code, CFLAGS = NULL, PKG_CPPFLAGS = NULL, PKG_LIBS = NULL, env = parent.frame(), 
+                    overwrite = "callme", verbosity = 0) {
   
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   # Sanity check code
@@ -83,6 +126,18 @@ compile <- function(code, PKG_CPPFLAGS = NULL, PKG_LDFLAGS = NULL, env = parent.
   stopifnot(nchar(code) > 0)
   stopifnot(!is.na(code))
   stopifnot(overwrite %in% c('all', 'callme', 'functions', 'none'))
+  
+  if (verbosity >= 4) {
+    if (!is.null(CFLAGS)) {
+      message("CFLAGS      : ", CFLAGS)
+    }
+    if (!is.null(PKG_CPPFLAGS)) {
+      message("PKG_CPPFLAGS: ", PKG_CPPFLAGS)
+    }
+    if (!is.null(PKG_LIBS)) {
+      message("PKG_LIBS    : ", PKG_LIBS)
+    }
+  }
   
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   # Create a new directory to work in.
@@ -126,18 +181,48 @@ compile <- function(code, PKG_CPPFLAGS = NULL, PKG_LDFLAGS = NULL, env = parent.
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   write(code, c_file)
   
-  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  # Write a 'Makevars' with the CPPFLAGS because as-far-as-I-know
-  # you cannot set CPPFLAGs when doing "R CMD SHLIB" 
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  # Set 'CFLAGS' in a temporary R_MAKEVARS_USER
+  # If CFLAGS is set here, then this **replaces** the default CFLAGS R 
+  # would use when compiling
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  makevars_user <- file.path(tmp_dir, "makevars-user.txt")
+  
+  if (!is.null(CFLAGS)) {
+    stopifnot(is.character(CFLAGS))
+    stopifnot(length(CFLAGS) == 1)
+    stopifnot(nchar(CFLAGS) > 0)
+    stopifnot(!is.na(CFLAGS))
+    CFLAGS <- paste0("CFLAGS=", CFLAGS)
+    writeLines(CFLAGS, makevars_user)
+  }
+  
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  # Write a 'Makevars' with the PKG_CPPFLAGS, PKG_LIBS
+  # AFAIK: you cannot set PKG_CPPFLAGS on the command line when doing "R CMD SHLIB" 
+  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  makevars <- c("")
+  
+  
   if (!is.null(PKG_CPPFLAGS)) {
     stopifnot(is.character(PKG_CPPFLAGS))
     stopifnot(length(PKG_CPPFLAGS) == 1)
     stopifnot(nchar(PKG_CPPFLAGS) > 0)
     stopifnot(!is.na(PKG_CPPFLAGS))
     PKG_CPPFLAGS <- paste0("PKG_CPPFLAGS=", PKG_CPPFLAGS)
-    writeLines(PKG_CPPFLAGS, "Makevars")
+    makevars <- c(makevars, PKG_CPPFLAGS)
   }
+  
+  if (!is.null(PKG_LIBS)) {
+    stopifnot(is.character(PKG_LIBS))
+    stopifnot(length(PKG_LIBS) == 1)
+    stopifnot(nchar(PKG_LIBS) > 0)
+    stopifnot(!is.na(PKG_LIBS))
+    PKG_LIBS <- paste0("PKG_LIBS=", PKG_LIBS)
+    makevars <- c(makevars, PKG_LIBS)
+  }
+  
+  writeLines(makevars, "Makevars")
   
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   # Setup stdout/stderr
@@ -152,15 +237,17 @@ compile <- function(code, PKG_CPPFLAGS = NULL, PKG_LDFLAGS = NULL, env = parent.
   
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   # Compile a shared library.
-  # LDFLAGs can be included at the end of a "R CMD SHLIB" call to add
-  # extra library search paths and link to libraries.
+  # Run with a temporary R_MAKEVARS_USER so that we an override
+  # R's default CFLAGS 
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  ret <- system2(
-    command = paste0(R.home(component = "bin"), "/R"), 
-    args    = paste("CMD SHLIB", basename(c_file), PKG_LDFLAGS),
-    stdout  = stdout,
-    stderr  = stdout
-  )
+  ret <- with_makevars_file_env(
+    R_MAKEVARS_USER = makevars_user, 
+    system2(
+      command = paste0(R.home(component = "bin"), "/R"), 
+      args    = paste("CMD SHLIB", basename(c_file)),
+      stdout  = stdout,
+      stderr  = stdout
+    ))
   
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   # Show errors
@@ -182,17 +269,17 @@ compile <- function(code, PKG_CPPFLAGS = NULL, PKG_LDFLAGS = NULL, env = parent.
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   dyn.load(dll_file)
   if (verbosity >= 2) {
-    cat("dll file: ", dll_file, "\n")
+    message("\nDLL file: ", dll_file, "\n")
   }
   
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   # Automatically generate some wrapper functions in a named list
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   func_list <- create_wrapper_functions(code, dll_file)
-
+  
   
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  # Assign functions into environment
+  # Assign wrapper functions into the nominated environment
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   if (!is.null(env)) {
     for (func_name in names(func_list)) {
@@ -208,7 +295,7 @@ compile <- function(code, PKG_CPPFLAGS = NULL, PKG_LDFLAGS = NULL, env = parent.
         } else {
           stop("compile(): Not allowed to overwrite existing variable with new wrapper function '", func_name, "'", call. = FALSE)
         }
-         
+        
         
       }
       assign(func_name, func_list[[func_name]], pos = env)